API documentation

Functions

CitableTeiReaders.readcitable — Function

Read from src a citable document identified by urn and cited by citation scheme citescheme.

readcitable(src, urn, citescheme)

Read from file fname a citable document identified by urn and cited by citation scheme citescheme.

readcitable(fname, urn, citescheme, rdr)

Read from url a citable document identified by urn and cited by citation scheme citescheme.

readcitable(url, urn, citescheme, rdr)

Implementation of readcitable for type TEIPoeticLine.

Implementation of readcitable for type TEIAnonblock.

Implementation of readcitable for type TEIDivAb.

Implementation of readcitable for type TEIDivLine.

Implementation of readcitable for type TEIDivDivDiv.

Implementation of readcitable for type TEIThreeDivsGrouped.

Implementation of readcitable for type DelimitedText.

CitableTeiReaders.ezxmlstring — Function

ezxmlstring(n)

Serialize an EzXML.Node to an XML string. If a node n was created by parsing an XML string, the output of ezxmlstring(n) should be XML equivalent to the source string.